oops

+++ source/common/http/http1/legacy_parser_impl.h
@@ -18,7 +18,7 @@ public:
   void resume() override;
   ParserStatus pause() override;
   int getErrno() override;
-  int statusCode() const override;
+  uint16_t statusCode() const override;
   int httpMajor() const override;
   int httpMinor() const override;
   uint64_t contentLength() const override;
